third_party/WebKit/LayoutTests/accessibility/scroll-to-make-visible-main-window-expected.txt

Tests that scrolling to make an element visible successfully scrolls the main window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Upper Target
5000-pixel box
Lower Target
5000-pixel box
PASS window.pageYOffset is 0
PASS window.pageYOffset >= minYOffset is true
PASS window.pageYOffset <= maxYOffset is true
PASS window.pageYOffset >= minYOffset is true
PASS window.pageYOffset <= maxYOffset is true
PASS successfullyParsed is true

TEST COMPLETE
